requirements.txt
webshooter
newspaper3k
pandas
ray
jupyter
boto3